+++ src/main/java/Duke.java
@@ -1,81 +1,15 @@
+import task.Task;
+import task.Todo;
+import task.Deadline;
+import task.Event;
+import exceptions.EmptyDescriptionException;
+import exceptions.UnknownCommandException;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
 
 public class Duke {
-
-    private static class EmptyDescriptionException extends Exception {
-        EmptyDescriptionException(String message) {
-            super(message);
-        }
-    }
-
-    private static class UnknownCommandException extends Exception {
-        UnknownCommandException(String message) {
-            super(message);
-        }
-    }
-
-    private static abstract class Task {
-        String description;
-        boolean isDone;
-
-        Task(String description) {
-            this.description = description;
-            this.isDone = false;
-        }
-
-        void markAsDone() {
-            this.isDone = true;
-        }
-
-        void unmark() {
-            this.isDone = false;
-        }
-
-        public abstract String toString();
-    }
-
-    private static class Todo extends Task {
-        Todo(String description) {
-            super(description);
-        }
-
-        @Override
-        public String toString() {
-            return "[T][" + (isDone ? "X" : " ") + "] " + description;
-        }
-    }
-
-    private static class Deadline extends Task {
-        String by;
-
-        Deadline(String description, String by) {
-            super(description);
-            this.by = by;
-        }
-
-        @Override
-        public String toString() {
-            return "[D][" + (isDone ? "X" : " ") + "] " + description + " (by: " + by + ")";
-        }
-    }
-
-    private static class Event extends Task {
-        String from;
-        String to;
-
-        Event(String description, String from, String to) {
-            super(description);
-            this.from = from;
-            this.to = to;
-        }
-
-        @Override
-        public String toString() {
-            return "[E][" + (isDone ? "X" : " ") + "] " + description + " (from: " + from + " to: " + to + ")";
-        }
-    }
 
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
@@ -133,23 +67,11 @@ public class Duke {
                         System.out.println("  " + tasks.get(tasks.size() - 1));
                         System.out.println("Now you have " + tasks.size() + " tasks in the list.");
                     }
-                } else if (input.startsWith("delete")) {
-                    int index = Integer.parseInt(input.split(" ")[1]) - 1;
-                    if (index >= 0 && index < tasks.size()) {
-                        System.out.println("Noted. I've removed this task:");
-                        System.out.println("   " + tasks.get(index));
-                        tasks.remove(index);
-                        System.out.println("Now you have " + tasks.size() + " tasks in the list.");
-                    } else {
-                        System.out.println("OOPS!!! No task found at the given index.");
-                    }
                 } else {
                     throw new UnknownCommandException("OOPS!!! I'm sorry, but I don't know what that means :-(");
                 }
             } catch (EmptyDescriptionException | UnknownCommandException e) {
                 System.out.println(e.getMessage());
-            } catch (NumberFormatException | IndexOutOfBoundsException e) {
-                System.out.println("OOPS!!! The index provided is not valid.");
             }
 
             System.out.println(line);
