+++ src/main/java/exceptions/EmptyDescriptionException.java
@@ -1,0 +1,8 @@
+package exceptions;
+
+public class EmptyDescriptionException extends Exception {
+    public EmptyDescriptionException(String message) {
+        super(message);
+    }
+}
+
