+++ src/main/java/task/Deadline.java
@@ -1,0 +1,16 @@
+package task;
+
+public class Deadline extends Task {
+    private String by;
+
+    public Deadline(String description, String by) {
+        super(description);
+        this.by = by;
+    }
+
+    @Override
+    public String toString() {
+        return "[D][" + (isDone ? "X" : " ") + "] " + description + " (by: " + by + ")";
+    }
+}
+
