+++ src/main/java/task/Todo.java
@@ -1,0 +1,13 @@
+package task;
+
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T][" + (isDone ? "X" : " ") + "] " + description;
+    }
+}
+
